refactor: add spring-jpa module
