+++ spring-java/spring-jpa/src/main/java/com/practice/spring_jpa/SpringJpaApplication.java
@@ -1,0 +1,13 @@
+package com.practice.spring_jpa;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringJpaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringJpaApplication.class, args);
+	}
+
+}
