+++ spring-java/spring-jpa/src/main/java/com/practice/spring_jpa/jpa/entity/Item.java
@@ -1,0 +1,29 @@
+package com.practice.spring_jpa.jpa.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor
+public class Item {
+  @Id @GeneratedValue
+  private Long itemId;
+
+  private String name;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "memberId")
+  private Member member;
+
+  public Item(String name, Member member) {
+    this.name = name;
+    this.member = member;
+  }
+}
