+++ spring-java/spring-jpa/src/test/java/com/practice/spring_jpa/SpringJpaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.practice.spring_jpa;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringJpaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
