--- spring-java/src/main/java/com/exercise/springjava/SpringJavaApplication.java
@@ -1,13 +1,0 @@
-package com.exercise.springjava;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringJavaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringJavaApplication.class, args);
-	}
-
-}
