--- spring-java/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.username=sa
-
-spring.jpa.show-sql=true
-spring.jpa.open-in-view=false
