--- spring-java/src/test/java/com/exercise/springjava/SpringJavaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.exercise.springjava;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringJavaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
